Adding windowItem.hotSpotsOffsetY, checkMovement splitting..
Adding item y-offsets called "windowItem.hotSpotOffsetY" so the window
is now fixed in height, based on its components, this might be worked on
later to allow resizing at some point but is required to register the
click handling on drawn items.

Splitted checkMovement into the functions activeMovement and
deactivateMovement

Some other code cleanups which mostly remove the resizing of items
inside the window provided height.