Modified pushBoundaries, adding window content area definition
pushBoundariens now takes the parameters of item, styleOffset (line,
rect) and coordOffset index for the drawing coordinates of the item.

Added design.contentScrollbar and design.contentScrollbarPlugY which
give the basic definition for the scrollbars on y axis.
Added a bunch of code to prepare rendering of window contents, this code
consists contentBoundaries, contentHotSpots, contentData and
contentActions and windowItem.contentArea

Renamed drawRenderItems to drawFolderItems

Minor code cleanups and added comments