Added window.contentHeight attribute, added random dummy item generation
The height of the content is now noted in the windowItem.contentHeight
variable, this is used to determine the possible scrolling boundaries
and is respected if scrolling is possible or not.

Dummy items now become generated in a range of 50 + 25 items mininum.

Dragging the scrollbarPlug will now not fire a legit click which did
trigger, on drag, functions like the maximize window control. Those have
to be pressed explicitly to trigger.

The contentHeight is now only calculated for file and folder items, but
not other window content.