Scrollable content area and related fixes
Added scrollable content for the app window, so controls are scrolled
with the canvas accordingly, mouse traps for cursor changes are scrolled
too and are hidden if out of the content areas.

Enabled clipping of controls and text display for input fields.

Minor bugfix so setWindowTitle action will work again on the
inputFields.

Added isMoveable boolean value to the window control items, items which
are set false to move will not be clipped and cannot be scrolled, this
is used for the scrollbars for now.