Added contentWidth to data window, added scrollbarX
The contentWidth property of a window object is now calculated on the
fly by the listed items width in total.

Added scrollbarX which allows horizontal scrolling of window contents.

Bugfixed some positioning errors when the scrollbars are inverted in the
window. Bugfixed content hiding and rendering in this cases too.